requirements.txt
faiss-cpu
plotly
numpy
langchain_openai
streamlit
scikit-learn
pandas
python-dotenv
langchain
langchain-community
openpyxl